fixes read_csv module name for import
