--- src/read-csv.py
@@ -1,14 +1,0 @@
-import csv
-import sys
-import os
-
-countylist = []
-newspaperlist = []
-def readFile(file):
-    with open(file, newline='') as csvfile:
-        reader = csv.DictReader(csvfile)
-        for row in reader:
-            countylist.append(row['counties'])
-            newspaperlist.append(row['localnewspapers'])
-    totallist = (countylist, newspaperlist)
-    return totallist
